+++ docs/conf.py
@@ -61,6 +61,8 @@ autodoc_typehints_format = 'short'
 
 # NOTE: The code below is required because of `defer_build=True` in dataclass definitions
 
-from dipdup.config import DipDupConfig
 from pydantic.dataclasses import rebuild_dataclass
+
+from dipdup.config import DipDupConfig
+
 rebuild_dataclass(DipDupConfig)  # type: ignore[arg-type]