+++ src/dipdup/hasura.py
@@ -81,11 +81,7 @@ query introspectionQuery($name: String!) {
     }
   }
 }
-""".replace(
-    '\n', ' '
-).replace(
-    '  ', ''
-)
+""".replace('\n', ' ').replace('  ', '')
 
 
 @dataclass